Add initial scripts for replay and Sirikata integration.
Add script that loads and replays tweets from a file, using HTTP
commands to Sirikata servers to create objects and tell them about new
tweets. Add an Emerson script which acts as a source of tweets in
Sirikata, instantiated by the replay script. Also add a client script
that starts up the visualization. Finally, add instructions for
setting up the components and replaying a log to the README.