Get grouped visualization working.
Make grouping work by reporting which group each tweet is part of and
only adjusting the size of the corresponding presence instead of
creating a new presence for every tweet. Change the visualization to
handle this, by using new event handler registration for changes to
visible properties and using them to adjust colors.