Create JSON "meshes" for tweet groups.
For each loaded tweet, tokenize it and create a "mesh" file containing
JSON of the text and tokens from each tweet in that small region. Save
them to disk and run an HTTP server in another thread to distribute
them. Also construct URLs which are now passed to the twitter_source
script so it can properly set a mesh URL. To support updates, a query
parameter is included (so the same filename can be used, but it should
require a separate request because the query parameter will be
different).